Module Sc_sys

module Env : sig ... end

module Ezcmd : sig ... end

Command line manager for external tools.

module File : sig ... end

A module for handling files. When a file value is created, it is ensured to exist. When creating a file value, one must make sure the directory in which the file is created exists, otherwise this module will raise an exception.

module Log : sig ... end

module Lwt_file : sig ... end

module Lwt_lazy : sig ... end

module Lwt_lock : sig ... end

module Lwt_memmap : sig ... end

Memory-mapping of files (LWT-ready)

module Lwt_task : sig ... end

module Lwt_watch : sig ... end

module Memmap : sig ... end

Memory-mapping of files

module Process : sig ... end

Safe process manager. Allows to start shell processes and monitor them.