Module Sc_sys.Exceptions

exception DryOperationError of string

Exception raised by Exceptions.err_if if called with dry=true. This is used when some operation that should be performed to properly setup a project is prevented due to the tool running in ``dry'' mode.

val log_error : Logs.src -> Logs.level -> exn -> unit

val err_if : ?dry:bool -> ('a, Stdlib.Format.formatter, unit, bool) Stdlib.format4 -> 'b

err_if ~dry format args... fails and raises Exn.DryOperationError with a string built using format and args..., or succeeds and returns false.