Module Sc_sys.File

A module for handling files. When a file value is created, it is ensured to exist. When creating a file value, one must make sure the directory in which the file is created exists, otherwise this module will raise an exception.

module TYPES : sig ... end

include module type of TYPES with type 'a file = 'a TYPES.file and type dir = TYPES.dir

type 'a file = 'a TYPES.file

The type of files. Having such a value ensures the file actually exists (unless it is removed by other means).

type dir = TYPES.dir

An alias for directories

Exceptions

exception INVALID_FILENAME of string

exception MISSING of {

file : 'a. 'a file;

}

exception UNEXPECTED of {

file : 'a. 'a file;

}

type 'a t = 'a file

Basics

val name : _ t -> string

Returns the file name given at creation.

val basename : _ t -> string

Returns the base name of the file.

val absname : _ t -> string

Returns the absolute path of the file.

val dirname : _ t -> string

Returns the directory name containing the file.

val print : Stdlib.Format.formatter -> _ t -> unit

Prints a file (given) name

val print_basename : Stdlib.Format.formatter -> _ t -> unit

Prints the basename of a file

val exists : _ t -> bool

val is_dir : _ t -> bool

val stat : _ t -> Unix.stats

val descriptor : _ t -> Unix.open_flag list -> Unix.file_perm -> Unix.file_descr

val check_suffix : _ t -> string -> bool

check_suffix file suff checks whether suff is a suffix for the basename of file.

val chop_suffix : _ t -> string -> string

val dir : _ t -> dir

Both dir f and parent f return the directory containing f.

val parent : _ t -> dir

Sanity checks

val fail_on_missing : _ t -> unit

val fail_on_missing_dir : dir -> unit

val fail_on_existing : _ t -> unit

Helpers for sanity checks

val missing : _ t -> exn

val unexpected : _ t -> exn

"Creation" functions

Not necessarily reflected on the filesystem: please use sanity checks when relevant.

val check_name : string -> unit

check_name filename raises INVALID_FILENAME if filename does not desribe a valid name for a file.

val from_name : string -> _ t

Creation functions below return file representations, and may perform various checks. The three first variants below do not actually perform any modification on the filesystem.

assume*

variants only check for invalid file names;

existing*

variants may additionally raise MISSING if the given file does not exist upon call;

not*

variants may additionally raise UNEXPECTED if the given file exists upon call, and MISSING if the parent directory does not exist;

create_empty*

acts like equivalent

not*

variants, except that the file is actually created;

mkdir*

variants creates a directory (recursively ascending to it parents if they do not exist --- equivalently to "mkdir -p"). May raise UNEXPECTED when a file is encountered in place of a directory;

val assume : string -> _ t

val assume_in : dir:dir -> string -> _ t

val assume_dir : string -> dir

val assume_dir_in : dir:dir -> string -> dir

val existing : string -> _ t

val existing_in : dir:dir -> string -> _ t

val existing_dir : string -> _ t

val existing_dir_in : dir:dir -> string -> _ t

val not_ : string -> _ t

val not_in : dir:dir -> string -> _ t

val create_empty : string -> _ t

val create_empty_in : dir:dir -> string -> _ t

val mkdir : string -> dir

val mkdir_in : dir:dir -> string -> dir

module PRETTY : sig ... end

"Touching"

(create if missing)

val touch : _ t -> unit

val touch_in : dir:dir -> string -> unit

val touch_dir : dir -> unit

Channels

val open_append : _ t -> Stdlib.out_channel

open_append file returns an out_channel for appending on file.

val safe_open_in : _ t -> (Stdlib.in_channel -> 'a) -> 'a

safe_open_in file cont Opens an in channel on `file` and applies `cont` on the channel. If any exception is raised out of `cont` or `cont` finishes its execution, the channel is closed.

val safe_open_out : _ t -> (Stdlib.out_channel -> 'a) -> 'a

safe_open_out file cont Opens an out channel on `file` and applies `cont` on the channel. If any exception is raised out of `cont`, the channel is closed.

val safe_open_append : _ t -> (Stdlib.out_channel -> 'a) -> 'a

safe_open_append file cont opens an out channel for appending on `file` and applies `cont` on the channel. If any exception is raised out of `cont`, the channel is closed.

Removal

val unlink : _ t -> unit

Removes a file. If the file is a directory, raises:

EPERM on POSIX compliant system

EISDIR on Linux >= 2.1.132

EACCESS on Windows

val rename_away : old_file:_ t -> new_basename:string -> _ t

rename_away ~old_file ~new_basename renames old_file by path_to_old_file/new_basename. Appends a time-dependent suffix to new_basename in order to make the new filename unique.

Returns the name of the replacement file.

Linking and copying

val copy_in : dir:dir -> 'a t -> 'a t

copy_in ~dir file Creates a copy of file in dir. Replaces any existing file from dir that has the same basename as file.

val link : ?dry:bool -> 'a t -> 'b t -> unit

val link_as : ?dry:bool -> 'a t -> string -> 'a t

val link_in_dir : ?dry:bool -> dir:dir -> 'a t -> 'a t

Searching among files or directories

val find_first_dir : dir list -> dir

find_first_dir dirnames returns the first element in dirnames that is the relative or absolute name of an existing directory, raises Not_found otherwise.

Digest/hash file or directory contents

val digest : _ t -> Stdlib.Digest.t

digest f assumes f is a regular file

val hash_files : _ t list -> string

val hash_dir : dir -> string

Syntax extensions

module Syntax : sig ... end