Module File.TYPES

type +_ file

The type of files. Having such a value ensures the file actually exists (unless it is removed by other means).

type dir = [ `dir ] file

An alias for directories

Exceptions

exception INVALID_FILENAME of string

exception MISSING of {

file : 'a. 'a file;

}

exception UNEXPECTED of {

file : 'a. 'a file;

}

type unix_error = {

context : string;
(*
describes when the error occured
*)

file_operands : abstract list;
(*
files that were operated on
*)

code : Unix.error;

func : string;

arg : string;
(*
as per Unix_error
*)

}

and abstract =

| A : _ file -> abstract

exception UNIX_ERROR of unix_error

exception DRY_OPERATION_ERROR of string

Exception raised by err_if if called with dry=true. This is used when some operation that should be performed to properly setup a project is prevented due to the tool running in ``dry'' mode.