Module Sc_sys.Lwt_file

Lwt-specific routines for handling files.

val descriptor : 'a File.t -> Unix.open_flag list -> Unix.file_perm -> Lwt_unix.file_descr Lwt.t

val exists : _ File.t -> bool Lwt.t

val misses : _ File.t -> bool Lwt.t

val stat : _ File.t -> Unix.stats Lwt.t

val link : ?dry:bool -> 'a File.t -> 'b File.t -> unit Lwt.t

val copy_in : dir:File.dir -> 'a File.t -> 'a File.t Lwt.t

val move_in : dir:File.dir -> _ File.t -> unit Lwt.t

val empty : _ File.t -> unit Lwt.t

val touch : _ File.t -> unit Lwt.t

val unlink : _ File.t -> unit Lwt.t

val newer_than : this:_ File.t -> _ File.t -> bool Lwt.t

Tells whether f2 is missing or it has a modification time older than this

val mkdir : string -> File.dir Lwt.t

mkdir dir Applies the command `mkdir -p dir` iff `dir` does not contain any of the following characters: ' ', '&', ' ', '|', ';'. Raises `Exceptions.Invalid_dir_name dir` otherwise

val touch_dir : File.dir -> unit Lwt.t

val dir_is_empty : File.dir -> bool Lwt.t

Returns true if the directory in argument is empty.

val files_of_dir : File.dir -> 'a File.t Lwt_stream.t

files_of_dir dir streams the names of every regular file (i.e, not directories) in dir.

val dirs_of_dir : File.dir -> 'a File.t Lwt_stream.t

dirs_of_dir dir streams the names of every sub-directory directly within dir.

val unlink_files_of_dir : File.dir -> unit Lwt.t

unlink_files_of_dir dir unlinks (i.e, removes) every regular file within dir.

val with_temp_file : ?flags:Unix.open_flag list -> ?perm:Unix.file_perm -> ?prefix:string -> ?suffix:string -> (_ File.t -> 'a Lwt.t) -> 'a Lwt.t

Equivalent to Lwt_io.with_temp_file, with a file instead of a pair of filename and an output channel.

Directory-wide locks

type dirlock_operation =

| Read

| ReadWrite

val with_lock_in : File.dir -> dirlock_operation -> (unit -> 'a Lwt.t) -> 'a Lwt.t

with_lock_in dir op f calls the promise f after having acquired a lock in directory dir (using a file `.lock` to hold locks). The lock is released whenever f is fulfilled or rejected. op indicates whether the lock is aqcuired only for reading files in dir or both reading and writing.

Management of symbolic links

val resymlink : to_dir:bool -> src:string -> dst:string -> unit Lwt.t

resymlink ~to_dir ~src ~dst creates or replaces a symbolic link dst to src. See Unix.symlink for restrictions, and semantics of to_dir.

val newsymlink : to_dir:bool -> src:string -> dstformat:(int -> string) -> string Lwt.t

newsymlink ~to_dir ~src ~dstformat creates a new symbolic link to src. The name of the link is returned; it is formed using dstformat with a sequence of increasing strictly positive Integers, and selecting the first name available for a symbolic link.

val redirect_symlinks_of : dir:File.dir -> src:string -> dst:File.dir -> unit Lwt.t

redirect_symlinks_of ~dir ~src ~dst updates all symbolic directories of dir that point to the entry basename src (within dir) so that it points to the entry whose basename is dst (within dir as well).

String-based manipulations

val read : _ File.t -> string Lwt.t

val write : _ File.t -> string -> unit Lwt.t

val digest : _ File.t -> Stdlib.Digest.t Lwt.t

val with_lines_of : file:_ File.t -> (string Lwt_stream.t -> 'a Lwt.t) -> 'a Lwt.t

Syntax extensions

module Syntax : sig ... end

Let-bindings for easy access to lwt I/O channels. See Lwt_io for operations available on such channels.