Module Lwt_file.Syntax

Let-bindings for easy access to lwt I/O channels. See Lwt_io for operations available on such channels.

val (let<*) : _ File.t -> (Lwt_io.input_channel -> 'a Lwt.t) -> 'a Lwt.t

let<* ic = file in e binds ic to an LWT input channel that reads from file in the context of e. ic is closed whenever e is fulfilled or rejected. May raise Unix.Unix_error.

val (let>*) : _ File.t -> (Lwt_io.output_channel -> 'a Lwt.t) -> 'a Lwt.t

let>* oc = file in e binds oc to an LWT output channel that writes to a new or emptied file in the context of e. oc is closed whenever e is fulfilled or rejected. May raise Unix.Unix_error.

val (let>*%) : _ File.t -> (Lwt_fmt.formatter -> 'a Lwt.t) -> 'a Lwt.t

Equivalent to let>*, but with a pretty-printing formatter.

val (let>*%!) : _ File.t -> (Stdlib.Format.formatter -> 'a Lwt.t) -> 'a Lwt.t

Equivalent to let>*%, but with a non-LWT pretty-printing formatter.

val (let>>*) : _ File.t -> (Lwt_io.output_channel -> 'a Lwt.t) -> 'a Lwt.t

let>>* oc = file in e binds oc to an LWT output channel that writes to the end of file in the context of e (file is created if it does not exist). oc is closed whenever e is fulfilled or rejected. May raise Unix.Unix_error.

val (let>>*%) : _ File.t -> (Lwt_fmt.formatter -> 'a Lwt.t) -> 'a Lwt.t

Equivalent to let>>*, but with a pretty-printing formatter.

val (let>>*%!) : _ File.t -> (Stdlib.Format.formatter -> 'a Lwt.t) -> 'a Lwt.t

Equivalent to let>>*%, but with a non-LWT pretty-printing formatter.