Module Lwt_watch.ASYNC

Asynchronous versions, where on_close may be called any number of times without waiting for previous non-terminated calls.

val monitor_dir : on_close:(_ File.t -> unit Lwt.t) -> File.dir -> (unit -> unit Lwt.t) Lwt.t

monitor_dir ~on_close dirname creates a thread that monitors file creations in directory dirname: on_close is subsequently called with any file f newly created in dirname.

The function promised must be called at most once to terminate the registration of f.

Please see limitations of monitor_dir for the case dirname does not exist upon call.