Module Process.TYPES

type 'a process

The type of processes that result in a value of type 'a

type command = string array

type failure = private Unix.process_status * command * string

Default type for errors raised when a spawned process errors out.

exception FAILED of failure

type redirection = [

| Lwt_process.redirection

| `Grab of stream_grabber

| `Log of line_logger

| `GrabNLog of stream_grabber * line_logger

]

Possible redirections for standard output and errors of processes spawned using exec and shell below. In addition to Lwt_process.redirection, `GrabLines f allows to pass the stream of lines to a function f, and `Log logfun passes each individual line to logfun.

and stream_grabber =

| Stream of string Lwt_stream.t -> unit Lwt.t

| Lines of string -> unit Lwt.t

| Push_lines of string option -> unit

Grabber for output streams (stdout or stderr). One may grab the stream as a whole using a promise or a mailbox variable, or its lines individially in sequence.

Note it is an error to assume that promises given via Stream or Lines are always executed before the sub-process terminates and its status is retrieved.

A useful pattern for grabbing output streams relies on the use of a mailbox variable as follows:

  let stderr_lines_mbox = Lwt_mvar.create_empty () in
  let* proc =
    Sc_sys.Process.exec ...
        ~stderr:(`Grab (Stream (Lwt_mvar.put stderr_lines_mbox)))
  and* stderr_lines =
    Lwt_mvar.take stderr_lines_mbox
  in
  ...

This pattern requires using a non-blocking spanwing function (i.e. exec or shell, but NOT shell_unit, shell_status, or exec_status) to launch the sub-process. If the underlying process does not terminate then the stream may be infinite.

Alternatively, use Push_lines to explicitly push the lines into a given stream. To use Push_lines, first create an empty stream, and then pass its push function to the process spawning function. Since the stream that is fed via Push_lines may only be closed during a join operation (unless an explicit push of None is performed), a join that happens in parallel with stream consumption is required to ensure: (i) no loss of any output line, and (ii) stream consumption does not hang on an empty channel. A correct pattern to achieve this (to, for instance, grab the standard error in case the sub-process errors out), is as follows:

  let stderr_lines, new_stderr_line = Lwt_stream.create () in
  Lwt.catch begin fun () ->
    Sc_sys.Process.join_lwt @@
    Sc_sys.Process.exec ...
      ~stderr:(`Grab (Push_lines new_stderr_line))
  end begin function
    | Sc_sys.Process.FAILED _ ->
      ... (* use stderr_lines as a usual stream *)
    | e ->
      Lwt.reraise e
  end

For ease of use according to this pattern, the function given to Push_lines may safely raise Lwt_stream.Closed, which is catched internally.

and line_logger = (string -> unit Lwt.t) Ez_logs.Types.log_lwt

Logger for individial lines.

type command_log_conditions = [

| `On_exec

| `On_exit

| `On_error

| `Always

]