Module Sc_sys

module Env : sig ... end

module Exceptions : sig ... end

module File : sig ... end

A module for handling files. When a file value is created, it is ensured to exist. When creating a file value, one must make sure the directory in which the file is created exists, otherwise this module will raise an exception.

module Log : sig ... end

module Log_lwt : sig ... end

module Lwt_file : sig ... end

module Lwt_lazy : sig ... end

module Lwt_watch : sig ... end

module Process : sig ... end

Safe process manager. Allows to start shell processes and monitor them.

module Version : sig ... end