Module Sc_sys.Lwt_file

Lwt-specific routines for handling files.

val descriptor : 'a File.t -> Unix.open_flag list -> Unix.file_perm -> Lwt_unix.file_descr Lwt.t

val exists : _ File.t -> bool Lwt.t

val misses : _ File.t -> bool Lwt.t

val stat : _ File.t -> Unix.stats Lwt.t

val link : ?dry:bool -> 'a File.t -> 'b File.t -> unit Lwt.t

val mkdir : string -> File.dir Lwt.t

mkdir dir Applies the command `mkdir -p dir` iff `dir` does not contain any of the following characters: ' ', '&', ' ', '|', ';'. Raises `Exceptions.Invalid_dir_name dir` otherwise

val dir_is_empty : File.dir -> bool Lwt.t

Returns true if the directory in argument is empty.

val files_of_dir : File.dir -> 'a File.t Lwt_stream.t

files_of_dir dir streams the names of all regular files (i.e., not directories) in dir.

val unlink_files_of_dir : File.dir -> unit Lwt.t

unlink_files_of_dir dir unlinks (i.e, removes) every regular file within dir.

Management of symbolic links

val resymlink : to_dir:bool -> src:string -> dst:string -> unit Lwt.t

resymlink ~to_dir ~src ~dst creates or replaces a symbolic link dst to src. See Unix.symlink for restrictions, and semantics of to_dir.

val newsymlink : to_dir:bool -> src:string -> dstformat:(int -> string) -> string Lwt.t

newsymlink ~to_dir ~src ~dstformat creates a new symbolic link to src. The name of the link is returned; it is formed using dstformat with a sequence of increasing strictly positive Integers, and selecting the first name available for a symbolic link.

val redirect_symlinks_of : dir:File.dir -> src:string -> dst:File.dir -> unit Lwt.t

redirect_symlinks_of ~dir ~src ~dst updates all symbolic directories of dir that point to the entry basename src (within dir) so that it points to the entry whose basename is dst (within dir as well).

Syntax extensions

module Syntax : sig ... end