Module Sc_sys.Process

Safe process manager. Allows to start shell processes and monitor them.

module TYPES : sig ... end

include module type of TYPES with type 'a process = 'a TYPES.process and type redirection = TYPES.redirection

type 'a process = 'a TYPES.process

The type of processes that result in a value of type 'a

type redirection = TYPES.redirection

Possible redirections for standard output and errors of processes spawned using exec and shell below. In addition to Lwt_process.redirection, `GrabLines f allows to pass the stream of lines to a function f, and `Log logfun passes each individual line to logfun.

type 'a t = 'a process

val exec : ?env:string array -> ?cwd:string -> ?stdin:Lwt_process.redirection -> ?stderr:redirection -> ?stdout:redirection -> ?timeout:float -> ?on_error:(Unix.process_status -> 'a Lwt.t) -> on_success:(unit -> 'a Lwt.t) -> Lwt_process.command -> 'a t

exec ?env ?cwd ?stdin ?stderr ?stdout ?timeout ?on_error ~on_success command spawns a new process using command. Standard input, output, and error streams are handled according to their respective optional arguments, where default values for stderr and stdout correspond to logging of each line using the log module attached to Process.

val shell : ?env:string array -> ?cwd:string -> ?stdin:Lwt_process.redirection -> ?stderr:redirection -> ?stdout:redirection -> ?timeout:float -> ?on_error:(Unix.process_status -> 'a Lwt.t) -> on_success:(unit -> 'a Lwt.t) -> string -> 'a t

Similar to exec, except command is launched in a shell.

val exec_status : ?stdout:redirection -> ?stderr:redirection -> Lwt_process.command -> bool Lwt.t

Returns with the process status whenever it terminates

val shell_status : ?stdout:redirection -> ?stderr:redirection -> string -> bool Lwt.t

Same as `exec_status` but with a shell command

val immediate : 'a -> 'a t

immediate res Is a process that returns `res` immediately

val join : 'a t -> 'a Lwt.t

Returns the promise associated with the process

val get_promise : 'a t -> 'a Lwt.t

val stdout : 'a t -> Lwt_io.input_channel option

Returns the standard output of the process

val stdout_lines : 'a t -> string Lwt_stream.t

val stdout_string : 'a t -> string Lwt.t

val stderr : 'a t -> Lwt_io.input_channel option

Returns the standard error of the process

val stderr_lines : 'a t -> string Lwt_stream.t

val stderr_string : 'a t -> string Lwt.t

val stdin : 'a t -> Lwt_io.output_channel option

Returns the standard input of the process

val stdin_lines : 'a t -> string Lwt_stream.t -> unit Lwt.t

val stdin_string : 'a t -> string -> unit Lwt.t

val print_unix_status : Stdlib.Format.formatter -> Unix.process_status -> unit

print_unix_status fmt status Prints the unix process status

val print_command : Stdlib.Format.formatter -> (string * string array) -> unit

Prints an lwt_process command

val kill : ?delay:float -> int -> 'a t -> unit Lwt.t

kill ?delay signum proc waits for delay seconds if this argument is provided, and then sends the signal signum to the process if it is still running.

val terminate : ?delay:float -> 'a t -> unit Lwt.t

terminate ?delay proc waits for delay seconds if this argument is provided, and then terminates the process if it is still running.