Module Process.TYPES

type 'a process

The type of processes that result in a value of type 'a

type redirection = [

| Lwt_process.redirection

| `GrabLines of string Lwt_stream.t -> unit Lwt.t

| `Log of (string -> unit Lwt.t) Logs_lwt.log

]

Possible redirections for standard output and errors of processes spawned using exec and shell below. In addition to Lwt_process.redirection, `GrabLines f allows to pass the stream of lines to a function f, and `Log logfun passes each individual line to logfun.